pyproject.toml
[tool.poetry]
name = "panos-cert-redistribution-checkerererer"
version = "0.1.0"
description = ""
authors = ["Calvin Remsburg <cremsburg.dev@gmail.com>"]
readme = "README.md"
packages = [{include = "panos_cert_redistribution_checkerererer"}]

[tool.poetry.dependencies]
python = "^3.11"
netmiko = "^4.3.0"
pan-os-python = "^1.11.0"
tabulate = "^0.9.0"
xmltodict = "^0.13.0"
dynaconf = "^3.2.4"


[tool.poetry.group.dev.dependencies]
ipdb = "^0.13.13"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
